requirements.txt
# Web backend
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.1

# Frontend
streamlit>=1.32.0

# LangChain + LangGraph
langchain>=0.1.20
langchain-community>=0.0.10
langgraph>=0.0.45
langchain-groq>=0.1.3

# Search tools
duckduckgo-search>=4.1.0

# HTTP & utilities
requests>=2.31.0
pydantic>=2.6.0
tenacity>=8.2.3
